+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - 2025-11-??
+
+### Added
+
+- Add 'localizeCountryNames' option to localize country names based on the app locale ([#57](https://github.com/juanjotorres90/ngx-material-intl-tel-input/issues/57)).
+
 ## [20.1.2] - 2025-10-08
 
 ### Fixed